Move golden summaries into command module
